Separate engine and document Fixes #256
