+++ crates/bitang/src/document/camera.rs
