+++ crates/bitang/src/document/chart.rs
