+++ crates/bitang/src/document/draw.rs
