+++ crates/bitang/src/document/generate_mip_levels.rs
