+++ crates/bitang/src/document/scene.rs
